Added vaadin GUI and a simulation

+++ src/main/java/se/panamach/services/gps/type/TimeDistanceVector.java
@@ -4,7 +4,7 @@ package se.panamach.services.gps.type;
 public class TimeDistanceVector extends Location {
 
 	// The estimated time of arrival to destination
-	public double estimatedArrivalTime;
+	public long estimatedArrivalTime;
 	
 	// The destination
 	public Location destination;
